HS-206 Restruct / rename
